Move Label from VariableStarViewer to ScatterPlotViewer
