+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/ScatterPlotViewer/components/Label/Label.jsx
